only Kernel imports sklearn and numpy now, the others reference kernels copy (took too long to import multiple times).
no longer uses standardscalar.

self.clusters is an expandable list that replaces the form and functionality of cluster1 and cluster2

clustertype is now an option to decide between kmean, dbscan, or none

stats is its own init option now, seperate from cluster type

setclusters renamed to kmeans, and dbscan function added. both cluster, depending on the clustertype.

dbscan only considers a cluster if it is larger than 1/100th the number of pixels present in the kernel

scatterplot option for dbscan with plot = true

showscatterplot works if either self.type = pixels or clustertype = kernels.
if clustertype = kernels, then showscatterplot now adds the kernelcenter points individually, and each point is the color its L,a,b values represent.

added functions HunterLabToXYZ, XYZToRGB, and HunterLabToRGB.

HunterLabToRGB has an optional normalized parameter that defaults to False. if set to true, it will return the RGB values normalized to be between[0,1].