If stats = True, then kernel, Cob, and repline now ignores the clusterstype
setpixels method can now calculate the mean without an additional iteration if calculatemean = True.
this allows faster mean calculation when you don't want all the stats.

checkdistance is now determined at the cob level when clustertype = kmeans.

cob also now holds a boolean property for segregating, which is determined in checkdistance

showscatterplot has an option closepreviousplot, which will leave the previously opened plot open if it is False. it also now sets the plot title to the cob name

repline now ignores the last pixel  of the last kernel of every cob, as it was throwing errors for some cobs (only partial RGB values were stored for some reason)

setsegregating method in repline  determines whether it is segregating or not when a repline has multiple cobs. it chooses based on whether majority segregating or not.

repline.kmeans now handles the situation that the replien is segregating or not, for single or multiple cobs

repline.showscatterplot works only for replines with oine cob, and it calls the cobs showscatterplot

removed repline.cluster property for nonsegregating replines. Now the cluster(s) are stored in repline.clusters regardless of segregating

test() in repline now handles numbers between 0 and 9999 inclusive